% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{aggregateBR}
\alias{aggregateBR}
\title{Aggregate Brazilain data by date, state and region}
\usage{
aggregateBR(
  data,
  by = c("estado", "regiao", "brasil", "state", "region", "brazil"),
  date = FALSE
)
}
\arguments{
\item{by}{areal level of aggregation (all brazilian territory, regions, states)}

\item{date}{logical (default is FALSE); if TRUE, then variables at each date are summed up.}

\item{language}{language; currently only portuguese and english available.}
}
\value{
tibble/data.frame with the aggregated data data.
}
\description{
Aggregate Brazilain data by date, state and region
}
\examples{
\dontrun{
library(readCovid19)

# portuguese:
portuguese <-  downloadBR("pt")
aggregateBR(data=portuguese, date=FALSE)
aggregateBR(data=portuguese, by="regiao", date=FALSE)
aggregateBR(data=portuguese, by="estado", date=FALSE)
dim(aggregateBR(data=portuguese, by="estado", date=TRUE))
dim(aggregateBR(data=portuguese, by="regiao", date=TRUE))

english:
english <- downloadBR("en")
aggregateBR(data=english, by="brazil", date=FALSE)
aggregateBR(data=english, by="region", date=FALSE)
aggregateBR(data=english, by="state", date=FALSE)
aggregateBR(data=english, by="state", date=TRUE)
aggregateBR(data=english, by="region", date=TRUE)

}

}
