% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{available_date}
\alias{available_date}
\alias{avdate}
\title{Function to seach for the most recent available data date}
\usage{
available_date(url, lagdays)
}
\arguments{
\item{url}{desired url for searching the data set}

\item{lagdays}{number of retrospective days to search for}
}
\value{
the most recent available data date.
}
\description{
The function searchs and identifies, among all available data sets hosted at a given repository, the up-to-date data set.
}
\note{
This function has been adapted from the code provided by Pedro Seiti Fujita.
}
