% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{downloadBR}
\alias{downloadBR}
\title{Function to download data (at Brazilian level) from the official Brazilian's repository}
\usage{
downloadBR(language = c("pt", "en"))
}
\arguments{
\item{language}{language; currently only portuguese and english available.}
}
\value{
tibble/data.frame with the downloaded data.
}
\description{
Function to download data (at Brazilian level) from the official Brazilian's repository
}
\examples{
\dontrun{
library(readCovid19)
brasil <- downloadBR()
brasil
}

}
