% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{downloadCovid19}
\alias{downloadCovid19}
\alias{downloadCovid}
\title{Function to download COVID-19 data from web repositories}
\usage{
downloadCovid19(url = c("brgov", "jhu"), language = c("en", "pt"))
}
\arguments{
\item{url}{data repository's url}

\item{language}{language; currently only portuguese and english available}
}
\value{
tibble/data.frame containing the downloaded data
}
\description{
This function downloads the pandemic COVID-19 data from two repositories: the the official Brazilian's repository mantained by the Brazilian Government (https://covid.saude.gov.br), which contains only data of the pandemia in Brazil at states and region levels, and the data from the Johns Hopkins University's repository (https://github.com/CSSEGISandData/COVID-19), which has been widely used all over the world as a reliable source of data information on the COVID-19 pandemia at a global level (countries and territories). For more details, please see our package's vignette.
}
\examples{
\dontrun{
library(readCovid19)
brazil <- downloadCovid19(language="pt", url="brgov")
brazil

world <- downloadCovid19(language="pt", url="jhu")
world
}

}
