% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{downloadWorld}
\alias{downloadWorld}
\alias{downloadBR}
\title{Function to download data (at world level) from the Johns Hopkins University's repository}
\usage{
downloadWorld(language = c("en", "pt"))
}
\arguments{
\item{language}{language; currently only portuguese and english available.}
}
\value{
tibble/data.frame with the downloaded data.
}
\description{
Function to download data (at world level) from the Johns Hopkins University's repository
}
\examples{
\dontrun{
library(readCovid19)
world <- downloadWorld()
world

# selecting data from Italy:
italy <- filter(world, local=="Italy")
}

}
