% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCovid19.R
\docType{package}
\name{readCovid19}
\alias{readCovid19}
\title{The 'readCovid19' package.}
\description{
R package to import CODID-19 pandemia data into R. The world level data is downloaded from the Johns Hopkins University's repository. Brazilian data is currently available on states level, and are downloaded from the official Brazilian's repository at https://covid.saude.gov.br/.
}
